Move drclusterConditionExpect to test utils
Also, add required client arguments, such that this can be reused
across various envs and tests run in those envs.
